requirements.txt
future~=1.0.0
opencv-python~=4.10.0.84
pyzbar~=0.1.9
requests~=2.32.3
pillow~=10.3.0